requirements.txt
Django==4.2
djangorestframework==3.17.1
python-dotenv==1.0.0